% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_compactness_constraint.R
\name{add_compactness_constraint}
\alias{add_compactness_constraint}
\title{Add constraint to limit compactness}
\usage{
add_compactness_constraint(x, max_diameter)
}
\arguments{
\item{max_diameter}{\code{numeric} Maximum diameter value.}
}
\description{
Add constraint to a restoration problem (\code{\link[=restore_problem]{restore_problem()}} object
to specify the compactness of a solution.
}
\details{
TODO. Provide details on how diameter is calculated and what sensible values
should be.
}
\examples{
# TODO

}
