% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_components_constraint.R
\name{add_components_constraint}
\alias{add_components_constraint}
\title{Add constraint to limit the number of connected components}
\usage{
add_components_constraint(problem, min_nb_components, max_nb_components)
}
\arguments{
\item{min_nb_components}{\code{integer} Minimum number of connected components.}

\item{max_nb_components}{\code{integer} Maximum number of connected components.}
}
\description{
Add constraint to a restoration problem (\code{\link[=restore_problem]{restore_problem()}} object
to specify the number of connected components that can be
present within a solution.
}
\details{
TODO. Provide details on what components are and what sensible values
should be.
}
\examples{
# TODO

}
