% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_max_iic_objective.R
\name{add_max_mesh_iic}
\alias{add_max_mesh_iic}
\title{Add an objective to maximize the integral index of connectivity}
\usage{
add_max_mesh_iic(x)
}
\value{
An updated restoration problem (\code{\link[=restore_problem]{restore_problem()}} object.
}
\description{
Specify that a restoration problem (\code{\link[=restore_problem]{restore_problem()}} should
the integral index of connectivity (IIC).
}
\details{
TODO. Provide details on what IIC is.
}
\examples{
# TODO

}
