% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_no_objective.R
\name{add_no_objective}
\alias{add_no_objective}
\title{Configure the solver to only satisfy the constraints, without optimization
objective}
\usage{
add_no_objective(problem)
}
\arguments{
\item{problem}{\code{\link[=restopt_problem]{restopt_problem()}} Restoration problem object.}
}
\value{
An updated restoration problem (\code{\link[=restopt_problem]{restopt_problem()}} object.
}
\description{
Specify that a restoration problem (\code{\link[=restopt_problem]{restopt_problem()}} should satisfy
the constraints without optimization objective.
}
\details{
Using set_no_objective() in a restopt problem, the solver will
return the first solution found satisfying the constraint, without any
optimization objective. This "no objective" setting is set by default
when creating a restopt problem.
}
\examples{
\dontrun{
# load data
habitat_data <- rast(
  system.file("extdata", "habitat.tif", package = "restoptr")
)

restorable_data <- rast(
  system.file("extdata", "restorable.tif", package = "restoptr")
)

# plot data
plot(rast(list(habitat_data, restorable_data)), nc = 2)

# create problem
p <- restopt_problem(
       existing_habitat = habitat_data,
       restorable_habitat = restorable_data
) \%>\% add_no_objective()

# print problem
print(p)

# Solve problem
s <- solve(p)
#' # plot solution
plot(
  x = s, main = "solution",
  col = c("#E5E5E5", "#ffffff", "#b2df8a", "#1f78b4")
)
}

}
