% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_restorable_constraint.R
\name{add_restorable_constraint}
\alias{add_restorable_constraint}
\title{Add constraint to specify the available amount of surface for restoration}
\usage{
add_restorable_constraint(
  problem,
  min_restore,
  max_restore,
  cell_area,
  min_proportion = 1
)
}
\arguments{
\item{min_restore}{\code{integer} Minimum allowed area to restore in the solution.}

\item{max_restore}{\code{integer} Maximum allowed area to restore in the solution}

\item{cell_area}{\code{integer} Total area of a cell.}

\item{min_proportion}{\code{float} Minimum habitat proportion to consider a cell
as restored.}
}
\description{
Add constraint to a restoration problem (\code{\link[=restore_problem]{restore_problem()}} object
to specify specify the available amount of surface for restoration
}
\details{
TODO
}
\examples{
# TODO

}
