% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_settings.R
\name{add_settings}
\alias{add_settings}
\title{Add settings}
\usage{
add_settings(problem, precision = 4, time_limit = 0)
}
\arguments{
\item{problem}{\code{\link[=restopt_problem]{restopt_problem()}} Restoration problem object.}

\item{precision}{\code{numeric} Precision for calculations.
Defaults to 4.}

\item{time_limit}{\code{integer} Maximum permitted run time for optimization
(seconds).
Defaults to 0.}
}
\description{
Add settings to a restoration problem (\code{\link[=restopt_problem]{restopt_problem()}} object
to customize the optimization procedure.
}
\examples{
\dontrun{TODO}

}
