% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restopt_problem.R
\name{get_restorable_habitat}
\alias{get_restorable_habitat}
\title{Retrieve the restorable habitat (aggregated) data.}
\usage{
get_restorable_habitat(problem)
}
\arguments{
\item{problem}{\code{\link[=restopt_problem]{restopt_problem()}} Restoration problem object.}
}
\value{
\code{\link[terra:rast]{terra::rast()}} The restorable habitat (aggregated) data.
}
\description{
Retrieve the restorable habitat (aggregated) data.
}
\examples{
\dontrun{
#' # load data
habitat_data <- rast(
  system.file("extdata", "habitat_hi_res.tif", package = "restoptr")
)

# create problem
p <- restopt_problem_2(
       existing_habitat = habitat_data,
       aggregation_factor = 4,
       habitat_threshold = 0.7
)

plot(get_restorable_habitat(problem))
}

}
