% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestoptProblem.R
\name{hasCompactnessConstraint}
\alias{hasCompactnessConstraint}
\title{Return TRUE if a compactness constraint was posted to the model}
\usage{
hasCompactnessConstraint(problem)
}
\value{
TRUE if a compactness constraint was posted to the model, otherwise FALSE
}
\description{
Return TRUE if a compactness constraint was posted to the model
}
