% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestoptProblem.R
\name{hasNbComponentsConstraint}
\alias{hasNbComponentsConstraint}
\title{Return TRUE if a number of components constraints was posted to the model}
\usage{
hasNbComponentsConstraint(problem)
}
\value{
TRUE if a number of components constraints was posted to the model, otherwise FALSE
}
\description{
Return TRUE if a number of components constraints was posted to the model
}
