% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestoptProblem.R
\name{hasRestorableConstraint}
\alias{hasRestorableConstraint}
\title{Return TRUE if a restorable area constraint was posted to the model}
\usage{
hasRestorableConstraint(problem)
}
\value{
TRUE if a restorable area constraint was posted to the model, otherwise FALSE
}
\description{
Return TRUE if a restorable area constraint was posted to the model
}
