% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{is_binary_raster}
\alias{is_binary_raster}
\title{Is a binary raster?}
\usage{
is_binary_raster(x)
}
\arguments{
\item{x}{\code{\link[terra:rast]{terra::rast()}} Raster object.}
}
\value{
\code{logical} value indicating if raster contains binary values.
}
\description{
Check if \code{\link[terra:rast]{terra::rast()}} raster object contains binary values.
}
