% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_java_available.R
\name{is_java_available}
\alias{is_java_available}
\title{Is Java is available?}
\usage{
is_java_available()
}
\value{
A \code{logical} (\code{TRUE} or \code{FALSE}) value indicating if Java is available
for usage.
}
\description{
The \pkg{restoptr} package uses Java to perform optimization procedures.
As such, it is important that Java is installed correctly when using this
package. This function verifies if Java is available on the system.
}
