% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestoptProblem.R
\name{maximizeIIC}
\alias{maximizeIIC}
\title{Find a solution maximizing the integral index of connectivity (IIC)}
\usage{
maximizeIIC(problem, precision, time_limit = 0)
}
\arguments{
\item{problem}{A RestoptProblem instance}

\item{precision}{precision for IIC calculation}

\item{time_limit}{time limit for solving (use 0 for no time limit)}
}
\description{
Find a solution maximizing the integral index of connectivity (IIC)
}
