% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestoptProblem.R
\name{maximizeMESH}
\alias{maximizeMESH}
\title{Find a solution maximizing the effective mesh size (MESH)}
\usage{
maximizeMESH(problem, precision, time_limit = 0)
}
\arguments{
\item{problem}{A RestoptProblem instance}

\item{precision}{precision for MESH calculation}

\item{time_limit}{time limit for solving (use 0 for no time limit)}
}
\description{
Find a solution maximizing the effective mesh size (MESH)
}
