% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestoptProblem.R
\name{postCompactnessConstraint}
\alias{postCompactnessConstraint}
\title{Post a maximum diameter constraint to the problem}
\usage{
postCompactnessConstraint(problem, max_diameter)
}
\arguments{
\item{problem}{A RestoptProblem instance}

\item{max_diameter}{maxmimum allowed diameter for the smallest enclosing circle of the solution}
}
\description{
Post a maximum diameter constraint to the problem
}
