% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestoptProblem.R
\name{postNbComponentsConstraint}
\alias{postNbComponentsConstraint}
\title{Post a number of connected components constraint to the problem}
\usage{
postNbComponentsConstraint(problem, min_nb_components, max_nb_components)
}
\arguments{
\item{problem}{A RestoptProblem instance}

\item{min_nb_components}{minimum number of connected components allowed in the solution}

\item{max_nb_components}{maximum number of connected components allowed in the solution}
}
\value{
the updated problem
}
\description{
Post a number of connected components constraint to the problem
}
