% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestoptProblem.R
\name{postRestorableConstraint}
\alias{postRestorableConstraint}
\title{Post a restorable area constraint to the problem}
\usage{
postRestorableConstraint(
  problem,
  min_restore,
  max_restore,
  cell_area,
  min_proportion
)
}
\arguments{
\item{problem}{A RestoptProblem instance}

\item{min_restore}{minimum allowed area to restore in the solution}

\item{max_restore}{maximum allowed area to restore in the solution}

\item{cell_area}{total cell area}

\item{min_proportion}{minimum habitat proportion to consider a cell as restored}
}
\description{
Post a restorable area constraint to the problem
}
