% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_vectorize.R
\name{preprocess_input_vect}
\alias{preprocess_input_vect}
\title{Restopr input vectorization-based preprocessing function.}
\usage{
preprocess_input_vect(
  habitat_data,
  non_habitat_value,
  cellsize,
  area_unit = "ha"
)
}
\description{
Restopr input vectorization-based preprocessing function.
}
