% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restopt_problem.R
\name{print.restopt_problem}
\alias{print.restopt_problem}
\title{Print a restoration optimization problem}
\usage{
\method{print}{restopt_problem}(x, ...)
}
\arguments{
\item{x}{\code{\link[=restore_problem]{restore_problem()}} Restoration problem object.}
}
\description{
Display information about a restoration
}
\examples{
TODO.
}
