% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restopt_problem.R
\name{print.RestoptProblem}
\alias{print.RestoptProblem}
\title{Print a restoration optimization problem}
\usage{
\method{print}{RestoptProblem}(x, ...)
}
\arguments{
\item{x}{\code{\link[=restopt_problem]{restopt_problem()}} Restoration problem object.}

\item{...}{Arguments not used.}
}
\description{
Display information about a restoration optimization problem.
}
\examples{
TODO.

}
