% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restopt_component.R
\name{restopt_component}
\alias{restopt_component}
\title{Restoration problem component}
\usage{
restopt_component(name, class, data = list(), post)
}
\arguments{
\item{name}{\code{character} Name of the component.}

\item{class}{\code{character} Vector of classes to identify the component.}

\item{data}{\code{list} Object containing parameters for adding the component.
Defaults to an empty list.}

\item{java}{\code{character} Name Java method for adding component.}

\item{args}{\code{character} Vector used to map the names of the elements in
\code{data} to the Java method.
Defaults to an empty character vector.}
}
\value{
An object.
}
\description{
Create a component to add to a restoration optimization problem.
}
\examples{
# TODO

}
