% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restopt_problem.R
\name{restopt_problem}
\alias{restopt_problem}
\title{Restoration optimization problem}
\usage{
restopt_problem(existing_habitat, restorable_habitat)
}
\arguments{
\item{existing_habitat}{\code{\link[terra:rast]{terra::rast()}} Raster object containing binary
values that indicate if each planning unit contains habitat or not.}

\item{restorable_habitat}{\code{\link[terra:rast]{terra::rast()}} Raster object containing
integer values that indicate the amount of habitat that can
be restored within each planning unit.}
}
\value{
A new restoration problem (\code{RestoptProblem}) object.
}
\description{
Create a new restoration optimization problem using data that describe
the spatial distribution of existing habitat and potential for
restoration.
}
\details{
TODO.
}
\examples{
# TODO

}
