% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestoptProblem.R
\name{RestoptProblem}
\alias{RestoptProblem}
\title{Restopt base problem constructor}
\usage{
RestoptProblem(
  habitat,
  habitat_path,
  restorable,
  restorable_path,
  accessible,
  accessible_path,
  accessible_value = 1
)
}
\arguments{
\item{habitat}{Habitat input raster object (either habitat_path or habitat must be given)}

\item{habitat_path}{Path to the habitat input raster file (either habitat_path or habitat must be given)}

\item{restorable}{Restorable input raster object (either restorable_path or restorable must be given)}

\item{restorable_path}{Path to the restorable input raster file}

\item{accessible}{Accessible input raster object (either accessible_path or accessible must be given)}

\item{accessible_path}{Path to the accessible input raster file}

\item{accessible_value}{Value of accessible areas in the accessible input raster}
}
\description{
Restopt base problem constructor
}
