% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{restoptr}
\alias{restoptr}
\title{restoptr: Interface to the 'Restopt' Ecological Restoration Planning Software}
\description{
Interface to the \code{restopt} ecological restoration planning software,
which is specifically designed to identify connected, accessible, and
compact areas for ecological restoration, with an emphasis on reducing
habitat fragmentation and increasing habitat connectivity.
}
