% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restopt_solution.R
\docType{class}
\name{RestoptSolution-class}
\alias{RestoptSolution-class}
\title{Define RestoptSolution class as inheriting SpatRaster}
\description{
Define RestoptSolution class as inheriting SpatRaster
}
