% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_compactness_constraint.R
\name{add_compactness_constraint}
\alias{add_compactness_constraint}
\title{Add constraint to limit compactness}
\usage{
add_compactness_constraint(problem, max_diameter)
}
\arguments{
\item{problem}{\code{\link[=restopt_problem]{restopt_problem()}} Restoration problem object.}

\item{max_diameter}{\code{numeric} Maximum diameter value.}
}
\description{
Add constraint to a restoration problem (\code{\link[=restopt_problem]{restopt_problem()}} object
to specify the compactness of a solution.
}
\details{
TODO. Provide details on how diameter is calculated and what sensible values
should be.
}
\examples{
\dontrun{TODO}

}
