% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_locked_out_constraint.R
\name{add_locked_out_constraint}
\alias{add_locked_out_constraint}
\title{Add locked out constraint}
\usage{
add_locked_out_constraint(problem, data)
}
\arguments{
\item{data}{\code{\link[terra:rast]{terra::rast()}} Raster object containing binary values
that indicate which planning units cannot be selected for any restoration.}
}
\description{
Add constraint to a restoration problem (\code{\link[=restopt_problem]{restopt_problem()}} object
to specify that certain planning units cannot be selected
for any restoration activities.
}
\details{
TODO.
}
\examples{
\dontrun{TODO}

}
