% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_locked_out_constraint.R
\name{add_locked_out_constraint}
\alias{add_locked_out_constraint}
\title{Add locked out constraint}
\usage{
add_locked_out_constraint(problem, data, raster_value = 1, lock_out = FALSE)
}
\arguments{
\item{data}{\code{\link[terra:rast]{terra::rast()}} Raster object containing binary values
that indicate which planning units cannot be selected for any restoration.}

\item{raster_value}{Raster value of cells to lock out (or not to lock out,
see \code{lock_out} parameter)}

\item{lock_out}{If TRUE cell with value \code{raster_value} in \code{data} are locked
out, otherwise cells with values different from \code{raster_value} are locked
out.}
}
\description{
Add constraint to a restoration problem (\code{\link[=restore_problem]{restore_problem()}} object
to specify that certain planning units cannot be selected
for any restoration activities.
}
\details{
TODO.
}
\examples{
# TODO

}
