% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_max_iic_objective.R
\name{add_max_iic_objective}
\alias{add_max_iic_objective}
\title{Add an objective to maximize the integral index of connectivity}
\usage{
add_max_iic_objective(problem)
}
\arguments{
\item{problem}{\code{\link[=restopt_problem]{restopt_problem()}} Restoration problem object.}
}
\value{
An updated restoration problem (\code{\link[=restopt_problem]{restopt_problem()}} object.
}
\description{
Specify that a restoration problem (\code{\link[=restopt_problem]{restopt_problem()}} should
the integral index of connectivity (IIC).
}
\details{
TODO. Provide details on what IIC is.
}
\examples{
# TODO

}
