% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_max_mesh_objective.R
\name{add_max_mesh_objective}
\alias{add_max_mesh_objective}
\title{Add an objective to maximize effective mesh size}
\usage{
add_max_mesh_objective(problem)
}
\arguments{
\item{problem}{\code{\link[=restopt_problem]{restopt_problem()}} Restoration problem object.}
}
\value{
An updated restoration problem (\code{\link[=restopt_problem]{restopt_problem()}} object.
}
\description{
Specify that a restoration problem (\code{\link[=restopt_problem]{restopt_problem()}} should
maximize effective mesh size.
}
\details{
TODO. Provide details on what effective mesh size is.
}
\examples{
\dontrun{TODO}

}
\references{
TODO. Citation for effective mesh size.
}
