% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restopt_solution.R
\name{get_metadata}
\alias{get_metadata}
\title{Restopt solution metadata}
\usage{
get_metadata(restopt_solution, area_unit = "ha")
}
\arguments{
\item{restopt_solution}{\code{\link[=restopt_solution]{restopt_solution()}} Restopt solution
to this solution.}

\item{area_unit}{Unit of the area ("ha" for hectares, "m" for square meters,
"km" for square kilometers)}
}
\value{
A list containing the characteristics of the restopt solution.
}
\description{
Return the metadata associated with a restopt solution, which contains
the characteristics of the solution. The unit for area characteristics
can be chosen among "ha" (hectares), "m" (square meters), "km" (square
kilometers), and "cells" (cells from the original input habitat raster).
Note that the solving time is expressed in seconds.
}
