% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RestoptProblem.R
\name{postNbComponentsConstraint}
\alias{postNbComponentsConstraint}
\title{Post a number of connected components constraint to the problem}
\usage{
postNbComponentsConstraint(problem, min_nbCC, max_nbCC)
}
\arguments{
\item{problem}{A RestoptProblem instance}

\item{min_nbCC}{minimum number of connected components allowed in the solution}

\item{max_nbCC}{maximum number of connected components allowed in the solution}
}
\description{
Post a number of connected components constraint to the problem
}
