% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{solve.RestoptProblem}
\alias{solve.RestoptProblem}
\title{Solve a restoration optimization problem}
\usage{
\method{solve}{RestoptProblem}(a, b, ...)
}
\arguments{
\item{a}{\code{\link[=restopt_problem]{restopt_problem()}} Restoration problem object.}

\item{b}{Argument not used.}

\item{...}{Unused arguments.}
}
\value{
A \code{\link[terra:rast]{terra::rast()}} object.
}
\description{
Solve a restoration optimization problem to generate a solution.
}
\examples{
\dontrun{TODO}

}
